backend/requirements.txt
Pillow==5.4.1
Flask==1.0.2
flask_cors==3.0.7
